[eclipse/xtext#1679] ported xtend to java
