+++ org.eclipse.xtext.xtext.bootstrap/.classpath
@@ -6,12 +6,6 @@
 			<attribute name="gradle_used_by_scope" value="mwe2"/>
 		</attributes>
 	</classpathentry>
-	<classpathentry kind="src" output="bin/mwe2" path="xtend-gen">
-		<attributes>
-			<attribute name="gradle_scope" value="mwe2"/>
-			<attribute name="gradle_used_by_scope" value="mwe2"/>
-		</attributes>
-	</classpathentry>
 	<classpathentry kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER/org.eclipse.jdt.internal.debug.ui.launcher.StandardVMType/JavaSE-1.8"/>
 	<classpathentry kind="con" path="org.eclipse.buildship.core.gradleclasspathcontainer"/>
 	<classpathentry kind="output" path="bin/default"/>
